# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Go Copilot API
  description: |
    Authentication and Role-Based Access Control (RBAC) API.

    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Rate Limiting
    Sensitive endpoints are rate-limited. When limits are exceeded, a 429 response is returned with a Retry-After header.

    ## Permissions
    Access to resources is controlled via RBAC. Users are assigned roles, which contain permissions.
    Permission format: `resource:action` (e.g., `users:read`, `roles:create`)

    ## Common Flow Examples

    ### 1. User Registration and Login
    ```bash
    # Register a new user
    curl -X POST http://localhost:8080/api/v1/auth/register \
      -H "Content-Type: application/json" \
      -d '{"email": "user@example.com", "password": "SecurePass123!", "full_name": "John Doe"}'

    # Login to get tokens
    curl -X POST http://localhost:8080/api/v1/auth/login \
      -H "Content-Type: application/json" \
      -d '{"email": "user@example.com", "password": "SecurePass123!"}'
    ```

    ### 2. Making Authenticated Requests
    ```bash
    # Get current user profile
    curl -X GET http://localhost:8080/api/v1/auth/me \
      -H "Authorization: Bearer <access_token>"

    # List users (requires users:list permission)
    curl -X GET http://localhost:8080/api/v1/users \
...
    ```

    ### 3. Refreshing Tokens
...

...

...

...
servers:
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List users
      description: Get paginated list of users
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, banned]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires users:list permission
    post:
      tags:
        - Users
      summary: Create user
      description: Create a new user (admin only)
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires users:create permission
        '409':
          description: Email already exists
...
  /users/{id}/roles/{roleId}:
    post:
      tags:
        - Users
      summary: Assign role to user
      description: Assign a specific role to a user
      operationId: assignRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RoleId'
      responses:
        '200':
          description: Role assigned
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires roles:assign permission
        '404':
          description: User or role not found
        '409':
          description: Role already assigned
    delete:
      tags:
        - Users
      summary: Revoke role from user
      description: Remove a specific role from a user
      operationId: revokeRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RoleId'
      responses:
        '204':
          description: Role revoked
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires roles:assign permission
        '404':
          description: User or role not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: User ID
    RoleId:
      name: roleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Role ID
...
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        avatar:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, banned]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        total:
          type: integer
          format: int64
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
    CreateUserRequest:
      type: object
      required:
        - email
        - password
        - full_name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        full_name:
          type: string
        role_ids:
          type: array
          items:
            type: string
            format: uuid
...